Add OrbitalLight and overhaul lighting/shaders
Introduce OrbitalLight (header + implementation) and register it in the project. Update World to expose light count and use lights in GLSLMaterial: send model, normal matrices, material/light uniforms (including linear attenuation and enable flag) and nLights. Revise GLSLProgram uniform names and texture binding paths. Rewrite shaders (vertex + fragment) to use material/light uniforms, normalize normals, compute lighting per light type, and switch shininess to float. Adjust Material/Object3D to use float shininess and default material colors, fix mesh paths, add new Feng_Shui mesh, and update main to create/add lights and load the Feng Shui object.

+++ mapi/Material.h
@@ -15,7 +15,7 @@ protected:
 	
 	glm::vec4 color;
 	
-	int shininess;
+	float shininess;
 	
 	bool 
 		light,
@@ -51,9 +51,9 @@ public:
 	
 	void setShadow(bool value) { this->shadow = value; }
 	
-	int getShininess() { return shininess; }
+	float getShininess() { return shininess; }
 	
-	void setShininess(int value) { this->shininess = value; }
+	void setShininess(float value) { this->shininess = value; }
 	
 	bool getReflection() { return reflection; }
